1.1-Patched
This is patched 1.1 both 1.1 and possibly 1.0 had an issue with spawndata. It should be resetting the spawndata and spawnpotentials yes it does that now but, the point is wasn't fully reset and could have caused major issues in the future.

This is do to my actual event handler no technical aspects a single line I really derped up implementing my own custom forge event. It was forge event simple and I screwed it up